requirements.txt
accelerate==0.23.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cmake==3.25.0
comm==0.1.3
cryptography==3.4.8
#dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diffusers #==0.21.2
distro==1.7.0
einops
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.9.0
fqdn==1.5.1
fsspec==2023.9.2
httplib2==0.20.2
huggingface-hub==0.17.2
idna==3.4
importlib-metadata==4.6.4
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
# jupyter-archive==3.3.4
# jupyter_client==8.3.0
# jupyter-contrib-core==0.4.2
# jupyter-contrib-nbextensions==0.7.0
# jupyter_core==5.3.1
# jupyter-events==0.6.3
# jupyter-highlight-selected-word==0.2.0
#jupyter-lsp==2.2.0
# jupyter-nbextensions-configurator==0.6.3
# jupyter_server==2.7.0
# jupyter_server_terminals==0.4.4
# jupyterlab==4.0.2
# jupyterlab-pygments==0.2.2
# jupyterlab_server==2.23.0
# jupyterlab-widgets==3.0.8
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==15.0.7
lxml==4.9.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==3.0.1
more-itertools==8.10.0
mpmath==1.2.1
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.4
notebook_shim==0.2.3
numpy
oauthlib==3.2.0
overrides==7.3.1
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip
platformdirs==3.8.1
prometheus-client==0.17.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
#PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
#python-apt==2.4.0+ubuntu1
pytorch-lightning
python-dateutil==2.8.2
python-json-logger==2.0.7
PyYAML
pyzmq==25.1.0
referencing==0.29.1
regex==2023.8.8
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.8.10
safetensors==0.3.3
SecretStorage==3.3.1
Send2Trash==1.8.2
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.11.1
terminado==0.17.1
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
# torch==2.0.1+cu118
# torchaudio==2.0.2+cu118
# torchvision==0.15.2+cu118
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2

tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
transformers
#triton==2.0.0
typing_extensions==4.4.0
uri-template==1.3.0
urllib3==1.26.13
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
wheel==0.40.0
widgetsnbextension==4.0.8
zipp==1.0.0
librosa
pydub